Converted Ints to Doubles for more precision

+++ Mathtastic/Structures/Point.cs
@@ -5,9 +5,9 @@ namespace Mathtastic.Structures
 {
     public class Point
     {
-        public int X { get; set; }
-        public int Y { get; set; }
-        public int Z { get; set; }
+        public double X { get; set; }
+        public double Y { get; set; }
+        public double Z { get; set; }
 
         public Point()
         {
